libft

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: cseguier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 16:47:55 by cseguier          #+#    #+#             */
-/*   Updated: 2018/11/15 18:57:39 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/11/28 17:47:27 by cseguier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memccpy(void *dest, void const *src, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i <= (int)n)
+	while (i <= n)
 	{
 		((char*)dest)[i] = ((char*)src)[i];
 		if (((char*)dest)[i] == (char)c)
