+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: cseguier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 13:30:43 by cseguier          #+#    #+#             */
-/*   Updated: 2018/11/15 15:21:59 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/11/28 17:33:51 by cseguier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char const *str)
+#include "libft.h"
+
+size_t	ft_strlen(char const *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
