+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: cseguier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/15 17:36:26 by cseguier          #+#    #+#             */
-/*   Updated: 2018/11/15 18:59:37 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/11/28 18:06:00 by cseguier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 char	*ft_strncat(char *dest, char const *src, size_t n)
 {
-	int	cpt;
-	int	i;
+	size_t	cpt;
+	size_t	i;
 
 	i = 0;
 	cpt = 0;
 	while (dest[cpt] != '\0')
 		cpt++;
-	while (src[i] != '\0' && (int)n > i)
+	while (src[i] != '\0' && n > i)
 	{
 		dest[cpt] = src[i];
 		i++;
