+++ ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: cseguier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/15 18:50:08 by cseguier          #+#    #+#             */
-/*   Updated: 2018/11/15 19:00:32 by cseguier         ###   ########.fr       */
+/*   Updated: 2018/11/28 18:01:09 by cseguier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 char	*ft_strnew(size_t t)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (!(str = (char*)ft_memalloc(sizeof(char) * (t + 1))))
 		return (NULL);
-	while (i <= (int)t)
+	while (i <= t)
 	{
 		str[i] = '\0';
 		i++;
